server/requirements.txt
flask
flask-cors
python-dotenv
PyMuPDF
google-generativeai
pillow
ag2[gemini]
pymongo
bcrypt
PyJWT
Flask-Bcrypt